main menu fix
Former-commit-id: 9fd65628705e7e121cf9e0b506aaeeeafcb0732a

+++ core/src/main/com/bodyconquest/screens/GameOverScreen.java
@@ -94,7 +94,7 @@ public class GameOverScreen extends AbstractGameScreen implements Screen {
         super.setRectangles();
         backBounds =
                 new Rectangle(
-                        BodyConquest.V_WIDTH / 2 - backButton.getWidth() / 2,
+                    (int) (BodyConquest.V_WIDTH  / 2 - backButton.getWidth() / 2),
                         50,
                         backButton.getWidth(),
                         backButton.getHeight());
